LibWeb: Split TransformationStyleValue out of StyleValue.{h,cpp}
